Extract ZodParsedType
